+++ lib/db/src/schema/index.ts
@@ -72,3 +72,4 @@ export * from "./goedkeuring";
 export * from "./compliance";
 export * from "./declaraties";
 export * from "./algemene-inkoop";
+export * from "./actiepunten";